+++ integration/package.go
@@ -1,0 +1,3 @@
+// This file doesn't have integration flag, and only exists to tell Go tooling the name of the package.
+
+package main