+++ pkg/querier/chunk_tar_test.go
@@ -5,6 +5,7 @@ import (
 	"compress/gzip"
 	"context"
 	"io"
+	"io/ioutil"
 	"math"
 	"os"
 	"strconv"
@@ -13,15 +14,17 @@ import (
 
 	"github.com/pkg/errors"
 	"github.com/prometheus/prometheus/promql"
+	"github.com/prometheus/prometheus/util/testutil"
 	"github.com/stretchr/testify/require"
 	"github.com/weaveworks/common/user"
 
 	"github.com/cortexproject/cortex/pkg/chunk"
 	"github.com/cortexproject/cortex/pkg/querier/batch"
+	"github.com/cortexproject/cortex/pkg/querier/chunkstore"
 	"github.com/cortexproject/cortex/pkg/util"
 )
 
-func getTarDataFromEnv(t testing.TB) (query string, from, through time.Time, step time.Duration, store ChunkStore) {
+func getTarDataFromEnv(t testing.TB) (query string, from, through time.Time, step time.Duration, store chunkstore.ChunkStore) {
 	var (
 		err            error
 		chunksFilename = os.Getenv("CHUNKS")
@@ -48,16 +51,21 @@ func getTarDataFromEnv(t testing.TB) (query string, from, through time.Time, ste
 	return query, from, through, step, &mockChunkStore{chunks}
 }
 
-func runRangeQuery(t testing.TB, query string, from, through time.Time, step time.Duration, store ChunkStore) {
+func runRangeQuery(t testing.TB, query string, from, through time.Time, step time.Duration, store chunkstore.ChunkStore) {
+	dir, err := ioutil.TempDir("", t.Name())
+	testutil.Ok(t, err)
+	defer os.RemoveAll(dir)
+	queryTracker := promql.NewActiveQueryTracker(dir, 1, util.Logger)
+
 	if len(query) == 0 || store == nil {
 		return
 	}
 	queryable := newChunkStoreQueryable(store, batch.NewChunkMergeIterator)
 	engine := promql.NewEngine(promql.EngineOpts{
-		Logger:        util.Logger,
-		MaxConcurrent: 1,
-		MaxSamples:    math.MaxInt32,
-		Timeout:       10 * time.Minute,
+		Logger:             util.Logger,
+		ActiveQueryTracker: queryTracker,
+		MaxSamples:         math.MaxInt32,
+		Timeout:            10 * time.Minute,
 	})
 	rangeQuery, err := engine.NewRangeQuery(queryable, query, from, through, step)
 	require.NoError(t, err)