+++ pkg/querier/querier.go
@@ -14,7 +14,9 @@ import (
 
 	"github.com/cortexproject/cortex/pkg/chunk"
 	"github.com/cortexproject/cortex/pkg/querier/batch"
+	"github.com/cortexproject/cortex/pkg/querier/chunkstore"
 	"github.com/cortexproject/cortex/pkg/querier/iterators"
+	"github.com/cortexproject/cortex/pkg/querier/lazyquery"
 	"github.com/cortexproject/cortex/pkg/util"
 )
 
@@ -35,6 +37,12 @@ type Config struct {
 	// Needs to be configured for subqueries to work as it is the default
 	// step if not specified.
 	DefaultEvaluationInterval time.Duration
+
+	// Directory for ActiveQueryTracker. If empty, ActiveQueryTracker will be disabled and MaxConcurrent will not be applied (!).
+	// ActiveQueryTracker logs queries that were active during the last crash, but logs them on the next startup.
+	// However, we need to use active query tracker, otherwise we cannot limit Max Concurrent queries in the PromQL
+	// engine.
+	ActiveQueryTrackerDir string `yaml:"active_query_tracker_dir"`
 
 	// For testing, to prevent re-registration of metrics in the promql engine.
 	metricsRegisterer prometheus.Registerer `yaml:"-"`
@@ -58,6 +66,7 @@ func (cfg *Config) RegisterFlags(f *flag.FlagSet) {
 	f.DurationVar(&cfg.QueryIngestersWithin, "querier.query-ingesters-within", 0, "Maximum lookback beyond which queries are not sent to ingester. 0 means all queries are sent to ingester.")
 	f.DurationVar(&cfg.DefaultEvaluationInterval, "querier.default-evaluation-interval", time.Minute, "The default evaluation interval or step size for subqueries.")
 	f.DurationVar(&cfg.QueryStoreAfter, "querier.query-store-after", 0, "The time after which a metric should only be queried from storage and not just ingesters. 0 means all queries are sent to store.")
+	f.StringVar(&cfg.ActiveQueryTrackerDir, "querier.active-query-tracker-dir", "./active-query-tracker", "Active query tracker monitors active queries, and writes them to the file in given directory. If Cortex discovers any queries in this log during startup, it will log them to the log file. Setting to empty value disables active query tracker, which also disables -querier.max-concurrent option.")
 	cfg.metricsRegisterer = prometheus.DefaultRegisterer
 }
 
@@ -74,12 +83,6 @@ func (cfg *Config) Validate() error {
 	return nil
 }
 
-// ChunkStore is the read-interface to the Chunk Store.  Made an interface here
-// to reduce package coupling.
-type ChunkStore interface {
-	Get(ctx context.Context, userID string, from, through model.Time, matchers ...*labels.Matcher) ([]chunk.Chunk, error)
-}
-
 func getChunksIteratorFunction(cfg Config) chunkIteratorFunc {
 	if cfg.BatchIterators {
 		return batch.NewChunkMergeIterator
@@ -89,7 +92,7 @@ func getChunksIteratorFunction(cfg Config) chunkIteratorFunc {
 	return mergeChunks
 }
 
-func NewChunkStoreQueryable(cfg Config, chunkStore ChunkStore) storage.Queryable {
+func NewChunkStoreQueryable(cfg Config, chunkStore chunkstore.ChunkStore) storage.Queryable {
 	return newChunkStoreQueryable(chunkStore, getChunksIteratorFunction(cfg))
 }
 
@@ -106,18 +109,28 @@ func New(cfg Config, distributor Distributor, storeQueryable storage.Queryable) 
 		if err != nil {
 			return nil, err
 		}
-		return newLazyQuerier(querier), nil
+		return lazyquery.NewLazyQuerier(querier), nil
 	})
 
 	promql.SetDefaultEvaluationInterval(cfg.DefaultEvaluationInterval)
 	engine := promql.NewEngine(promql.EngineOpts{
-		Logger:        util.Logger,
-		Reg:           cfg.metricsRegisterer,
-		MaxConcurrent: cfg.MaxConcurrent,
-		MaxSamples:    cfg.MaxSamples,
-		Timeout:       cfg.Timeout,
+		Logger:             util.Logger,
+		Reg:                cfg.metricsRegisterer,
+		ActiveQueryTracker: createActiveQueryTracker(cfg),
+		MaxSamples:         cfg.MaxSamples,
+		Timeout:            cfg.Timeout,
 	})
 	return lazyQueryable, engine
+}
+
+func createActiveQueryTracker(cfg Config) *promql.ActiveQueryTracker {
+	dir := cfg.ActiveQueryTrackerDir
+
+	if dir != "" {
+		return promql.NewActiveQueryTracker(dir, cfg.MaxConcurrent, util.Logger)
+	}
+
+	return nil
 }
 
 // NewQueryable creates a new Queryable for cortex.
@@ -169,8 +182,8 @@ type querier struct {
 	mint, maxt  int64
 }
 
-// Select implements storage.Querier.
-func (q querier) Select(sp *storage.SelectParams, matchers ...*labels.Matcher) (storage.SeriesSet, storage.Warnings, error) {
+// SelectSorted implements storage.Querier.
+func (q querier) SelectSorted(sp *storage.SelectParams, matchers ...*labels.Matcher) (storage.SeriesSet, storage.Warnings, error) {
 	// Kludge: Prometheus passes nil SelectParams if it is doing a 'series' operation,
 	// which needs only metadata. Here we expect that metadataQuerier querier will handle that.
 	// In Cortex it is not feasible to query entire history (with no mint/maxt), so we only ask ingesters and skip
@@ -210,7 +223,13 @@ func (q querier) Select(sp *storage.SelectParams, matchers ...*labels.Matcher) (
 
 	// we have all the sets from different sources (chunk from store, chunks from ingesters,
 	// time series from store and time series from ingesters).
+	// mergeSeriesSets will return sorted set.
 	return q.mergeSeriesSets(result), nil, nil
+}
+
+// Select implements storage.Querier.
+func (q querier) Select(sp *storage.SelectParams, matchers ...*labels.Matcher) (storage.SeriesSet, storage.Warnings, error) {
+	return q.SelectSorted(sp, matchers...)
 }
 
 // LabelsValue implements storage.Querier.
@@ -239,7 +258,7 @@ func (q querier) mergeSeriesSets(sets []storage.SeriesSet) storage.SeriesSet {
 			// If there is error, we better report it.
 			err := set.Err()
 			if err != nil {
-				otherSets = append(otherSets, errSeriesSet{err: err})
+				otherSets = append(otherSets, lazyquery.NewErrSeriesSet(err))
 			}
 			continue
 		}