+++ pkg/ruler/rules/store.go
@@ -7,7 +7,7 @@ import (
 	"github.com/cortexproject/cortex/pkg/configs"
 	"github.com/cortexproject/cortex/pkg/configs/client"
 
-	"github.com/prometheus/prometheus/pkg/rulefmt"
+	legacy_rulefmt "github.com/cortexproject/cortex/pkg/ruler/legacy_rulefmt"
 )
 
 var (
@@ -29,11 +29,11 @@ type RuleGroupList []*RuleGroupDesc
 
 // Formatted returns the rule group list as a set of formatted rule groups mapped
 // by namespace
-func (l RuleGroupList) Formatted() map[string][]rulefmt.RuleGroup {
-	ruleMap := map[string][]rulefmt.RuleGroup{}
+func (l RuleGroupList) Formatted() map[string][]legacy_rulefmt.RuleGroup {
+	ruleMap := map[string][]legacy_rulefmt.RuleGroup{}
 	for _, g := range l {
 		if _, exists := ruleMap[g.Namespace]; !exists {
-			ruleMap[g.Namespace] = []rulefmt.RuleGroup{FromProto(g)}
+			ruleMap[g.Namespace] = []legacy_rulefmt.RuleGroup{FromProto(g)}
 			continue
 		}
 		ruleMap[g.Namespace] = append(ruleMap[g.Namespace], FromProto(g))